add in data type cast to ensure stringent compilers are happy.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@5086 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/EdkCompatibilityPkg/Compatibility/FrameworkHiiToUefiHiiThunk/Forms.c
@@ -185,7 +185,7 @@ AppendToUpdateBuffer (
   }
   
   CopyMem (UefiData->Data + UefiData->Offset, OpCodeBuf, BufSize);
-  UefiData->Offset += BufSize;
+  UefiData->Offset += (UINT32) BufSize;
 
   return EFI_SUCCESS;
 }
